containers/tensorflow_models/requirements.txt
Cython>=0.19.2
numpy>=1.7.1
matplotlib>=1.3.1
Pillow>=2.3.0
six>=1.1.0
Flask>=0.12.2
Flask-Cors>=3.0.3
tqdm>=4.17.1
scikit-learn>=0.19.0
pandas>=0.20.3
